Adding more options to H5Writer
