+++ src/embkit/preprocessing/__init__.py
@@ -5,4 +5,4 @@ Proprocessing
 
 from sklearn.preprocessing import RobustScaler, MinMaxScaler
 from .normalize import quantile_max_norm, exp_max_norm, ExpMinMaxScaler, get_dataset_nonzero_mask
-from .dataset import DatasetMask
+from .dataset import DatasetMask, batch_iterable
